% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genField.R
\name{alternatives}
\alias{alternatives}
\title{Generate alternatives}
\usage{
alternatives(n, s, type = 1L, middle, delta = 0.15, distFun = dist)
}
\arguments{
\item{n}{dimensions of the requested random field. Numeric vector.}

\item{s}{parameter for the size of the blocks, 0.5 < s < 1, block length \eqn{l_n = n^s}.}

\item{type}{alternative type (integer or character). See "Details".}

\item{middle, delta, distFun}{parameters for alternative 4L. See "Details".}
}
\value{
\item{Types 1, 2, 4, 5}{A vector of indices.}
\item{Type 3}{A numeric (n x n) matrix containing the heights of the shifts.}
}
\description{
Generate alternatives
}
\details{
Alternatives:
\itemize{
\item 1L or "1a": exactly one block is shifted
\item "1b": size of the shift region is one block, but the shift region lies in two blocks
\item "1c": size of the shift region is one block, but the shift region lies in four blocks
\item 2L: exactly half of the data is shifted
\item 3L: there is a steady increase from left to right
\item 4L: a "circle" including everything within \code{distFun} \code{delta} from \code{middle} is shifted
\item 5L: for demonstration purposes: the field is divided into 10 "columns". Every other column is shifted
}
}
\examples{
alternatives(c(50, 50), 0.6)
alternatives(c(50, 50), 0.6, type = 4L, middle = c(10, 10))
alternatives(c(50, 50), 0.6, type = 3L)


}
\seealso{
\link{genField}
}
