% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocov}
\alias{autocov}
\title{Autocorrelation matrix}
\usage{
autocov(X, b, direction = 0L)
}
\arguments{
\item{X}{numeric matrix,}

\item{b}{numeric vector containing exactly two values: the bandwidths for the row- reps. column-wise estimation.}
}
\description{
Autocorrelation matrix
}
