% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockpValue.R
\name{block.pValue}
\alias{block.pValue}
\title{Block test p-value}
\usage{
block.pValue(tn, fun = "gmd")
}
\arguments{
\item{tn}{test statistic}

\item{fun}{Character string; one of "gmd" (default), "var", "jb", "ks", "grubbs", "ANOVA" (see \code{\link{block.stat}} for details).}
}
\value{
A numeric value between 0 and 1.
}
\description{
Returns the p-value of a test statistic according to the block test for structural changes.
}
