% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_pValue.R
\name{block_pValue}
\alias{block_pValue}
\title{Block test p-value}
\usage{
block_pValue(tn, fun = "gmd")
}
\arguments{
\item{tn}{test statistic}

\item{fun}{Character string; one of "gmd" (default), "var", "jb", "grubbs", "ANOVA", "ad", "sw" (see \code{\link{block_stat}} for details).}
}
\value{
A numeric value between 0 and 1.
}
\description{
Returns the p-value of a test statistic according to the block test for structural changes.
}
