% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crit.grubbs}
\alias{crit.grubbs}
\title{???????}
\usage{
crit.grubbs(n, alpha)
}
\arguments{
\item{n}{sample size; positive numeric value.}

\item{alpha}{significance level; between 0 and 1.}
}
\description{
???????
}
