% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{cutoff}
\alias{cutoff}
\title{Cut-Off}
\usage{
cutoff(X, ln)
}
\arguments{
\item{X}{data, numeric vector or matrix}

\item{ln}{l block length. Integer vector of length 1 or 2, depending on the number of dimensions of X, with strictly positive entries.}
}
\value{
A numeric vector of length \code{floor(n[1] / l[1]) * floor(n[2] / l[2])}, \code{n = dim(X)}.
}
\description{
This function returns a vector of the cut block means for a given time series or random field X.
}
\details{
For each block of size \code{ln[1]} \eqn{\cdot} \code{ln[2]}, the \code{ln[1]}-th row and the \code{ln[2]}-th column are deleted. Then, the mean is taken over the remaining data points.
}
\examples{
X <- genField(c(40, 80), H = 100, type = 2)
M <- cutoff(X, l = c(4, 4))

plot(X)
image(M)

}
