% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorr.R
\name{decorr}
\alias{decorr}
\title{De-correlation}
\usage{
decorr(X, lags, method = 1, separable = FALSE)
}
\arguments{
\item{X}{Random Field or time series}

\item{lags}{}

\item{method}{}

\item{separable}{}
}
\description{
De-correlation
}
