% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genTheta}
\alias{genTheta}
\title{Dependency Matrix Theta}
\usage{
genTheta(q, param)
}
\arguments{
\item{q}{model order (integer).}

\item{param}{MA parameter (numeric value between 0 and 1).}
}
\value{
A matrix of size (2q + 1) x (2q + 1).
}
\description{
This function generates a symmetric dependency matrix \eqn{\Theta} of a
specific type of spatial MA(q) model.
}
\details{
Symmetric spatial MA(q) model with parameter \code{param} \eqn{= \theta}:
\deqn{Y_{ij} = \sum_{k = -q}^q \sum_{l = -q}^q \theta^{|k - q - 1| + |l - q - 1|} \varepsilon_{kl}.}
\eqn{\{\theta^{|k - q - 1| + |l - q - 1|}\}_{kl} = \Theta}.
}
\examples{
genTheta(1, 0.2)

}
