% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GMD}
\alias{GMD}
\title{Gini's mean difference}
\usage{
GMD(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
A numeric value.
}
\description{
Calculates Gini's mean difference of a given vector \code{x}.
}
\examples{
x <- rnorm(100)
GMD(x)

}
