% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{skewness}
\alias{skewness}
\alias{kurtosis}
\title{Skewness and kurtosis}
\usage{
skewness(x)

kurtosis(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
A numeric value.
}
\description{
Compute the skewness and the kurtosis of the data vector \code{x}.
}
\examples{
skewness(rnorm(100))
skewness(rexp(100, 2))

kurtosis(rnorm(100))
kurtosis(rt(100, 5))
}
