% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lrv}
\alias{lrv}
\title{Long run variance}
\usage{
lrv(X, b = 1L)
}
\arguments{
\item{X}{numeric matrix,}

\item{b}{numeric vector containing exactly two values: the bandwidths for the row- reps. column-wise estimation.}
}
\value{
A numeric value.
}
\description{
Estimates the long run variance of a 2-dimensional matrix \code{X} using kernel
density estimation and the Tukey-Hanning kernel function.
}
\details{
??????? Tukey-Hanning kernel ????????
}
\examples{
X1 <- genField(c(50, 50), Phi = genPhi(1, 0.4))
b <- lrvBandwidth(X1, 1/3, 2/3)
lrv(X1, b)

Phi <- matrix(c(0.08, 0.1, 0.08, 0.8, 1, 0.8, 0.08, 0.1, 0.08), ncol = 3)
X2 <- genField(c(50, 50), Phi = Phi)
b <- lrvBandwidth(X2, 1/3, 2/3)
lrv(X2, b)

}
