% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrvBandwidth.R
\name{lrvBandwidth}
\alias{lrvBandwidth}
\title{Long run variance estimation bandwidth(s)}
\usage{
lrvBandwidth(X, p1, p2)
}
\arguments{
\item{X}{numeric matrix.}

\item{p1, p2}{exponents for sample size n resp. estimated dependency, between 0 and 1.}
}
\value{
A numeric vector containing two elements: the first value is the bandwidth for the row-wise
and the second one for the column-wise estimation.
}
\description{
Calculate optimal bandwidths for the long run variance (lrv) estimation of a 2-dim random field.
}
\details{
.................
}
\examples{
X1 <- genField(c(50, 50), Phi = genPhi(1, 0.4))
lrvBandwidth(X1, 1/3, 2/3)

Phi <- matrix(c(0.08, 0.1, 0.08, 0.8, 1, 0.8, 0.08, 0.1, 0.08), ncol = 3)
X2 <- genField(c(50, 50), Phi = Phi)
lrvBandwidth(X2, 1/3, 2/3)

}
