% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plot.RandomField}
\alias{plot.RandomField}
\title{Plot random field}
\usage{
\method{plot}{RandomField}(x, main = "", colors, name = "value", ...)
}
\arguments{
\item{x}{\code{RandomField} object}

\item{main}{title to the plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Plot random field
}
\examples{
plot(genField(c(50, 50)))
plot(genField(c(50, 50), type = 2))

}
