% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{pValue}
\alias{pValue}
\title{p-value}
\usage{
pValue(tn, fun = "gmd")
}
\arguments{
\item{tn}{test statistic}

\item{fun}{Character string; one of "gmd" (default), "var", "jb", "ks", "grubbs", "ANOVA" (see \code{\link{Tn}} for details).}
}
\value{
A numeric value between 0 and 1.
}
\description{
Returns the p-value of a test statistic according to the :::TODO: TESTNAME EINFUEGEN!!::: test.
}
