% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SChangeBlock-package.R
\docType{package}
\name{SChangeBlock-package}
\alias{SChangeBlock}
\alias{SChangeBlock-package}
\title{SChangeBlock: Spatial Structural change detection for time series and random fields by an analysis of variability between blocks of observations}
\description{
------Package Description--------
}
\author{
\strong{Maintainer}: Sheila Goerz \email{sheila.goerz@tu-dortmund.de}

}
\keyword{internal}
