% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sOpt}
\alias{sOpt}
\title{Optimal parameter s}
\usage{
sOpt(n, s = 0.6)
}
\arguments{
\item{n}{Sample size(s), numeric (vector).}

\item{s}{Desired exponent, \eqn{0.5 \leq s \leq 1}.}
}
\value{
Numeric Vector of the optimal exponents.
}
\description{
Calculates the best parameter for a given approximation s.
}
