% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCP-package.R
\docType{package}
\name{spatialCP-package}
\alias{spatialCP}
\alias{spatialCP-package}
\title{spatialCP: Structural change detection for time series and random fields by an analysis of variability between blocks of observations}
\description{
------Package Description--------
}
\author{
\strong{Maintainer}: Sheila Goerz \email{sheila.goerz@tu-dortmund.de}

}
\keyword{internal}
