% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sSizes}
\alias{sSizes}
\title{Optimal sizes ..................}
\usage{
sSizes(n, lower = 0.5, upper = 1, step = 0.1)
}
\arguments{
\item{n}{integer value.}

\item{lower, upper}{lower and upper search border, between 0 and 1.}

\item{step}{size of the step for the search, between 0 and 1.}
}
\value{
A data frame containing
\item{n}{the given sample size}
\item{s}{the exponent in question}
\item{ln}{the resulting block length}
\item{bn}{the corresponding number of block}
\item{ln.bn}{block length times number of blocks}
\item{diff}{difference between the given sample size and the number of observations covered by the blocks}
}
\description{
................................
}
\details{
........
}
\examples{
sSizes(50)
sSizes(50, 0.6, 0.8, 0.01)

}
