% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genField.R
\name{genField}
\alias{genField}
\title{Generate random fields}
\usage{
genField(
  n,
  distr = rnorm,
  type = 0L,
  H = 100,
  Theta = NULL,
  q = NULL,
  param = NULL,
  ...
)
}
\arguments{
\item{n}{dimensions of the requested random field. Numeric vector.}

\item{distr}{function specifying the error distribution.}

\item{type}{alternative type (integer or character). See \code{\link{alternatives}}.}

\item{H}{height of the location shift (numeric value).}

\item{Theta}{matrix specifying the dependency between observations of a 2x2 random field.
Has to be a 2-dim. matrix where there is an odd number of entries in both dimensions.
Explanations are given under "Details".}

\item{q}{dependency parameter for the model order of the MA field (integer > 0).}

\item{param}{dependency parameters for the model parameters of the MA field (numeric vector.)
Both \code{q} and \code{param} are ignored if \code{Theta} is supplied.
A dependency matrix is generated using \code{\link{genTheta}}.}

\item{...}{additional arguments for the generation of \code{\link{alternatives}}.}
}
\value{
The function returns a matrix of dimension \code{n} that has the \code{\link{class}} "RandomField".
}
\description{
Generate random fields on a regular grid. Dependency can be modeled to some extend.
}
\details{
The dependent random field is generated as follows:
Denote with \eqn{(e_{ij})} the error matrix from which the random field \eqn{(Y_{ij})}
(under the hypothesis, i.e. without any location shift) is generated. Assume the given dependency
matrix \code{Theta} has the dimension \eqn{d_1 \times d_2}, i.e. \code{Theta} = \eqn{(\theta)_{1\leq i \leq d_1 \atop 1 \leq j \leq d_2}}.
We define \eqn{p_1 = \lfloor d_1 / 2 \rfloor} and \eqn{p_2 = \lfloor d_2 / 2 \rfloor}.
Then for each observation \eqn{Y_{ij}, \; i = 1, ...,} \code{n[1]}, \eqn{j = 1, ..., } \code{n[2]}
we compute
\deqn{Y_{ij} = \sum_{h = -p_1}^{p_1} \sum_{k = -p_2}^{p_2} e_{i + h, j + k} \, \cdot \, \theta_{h + p_1 + 1, k + p_2 + 1}.}
That is, for each observation, we matrix-multiply the observation and the corresponding surrounding
ones with the dependency matrix \code{Theta}.
}
\examples{
genField(c(50, 50))
genField(c(50, 50), type = 3, Theta = genTheta(1, 0.2))

}
\seealso{
\link{alternatives}, \link{genTheta}
}
