% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mu}
\alias{Mu}
\title{Mu}
\usage{
Mu(x, group = NULL, l = NULL)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{group}{strictly positive integer vector or matrix indicating the group (or block) of the corresponding observation in X.
Overwrites \code{l} if specified.}

\item{l}{block length. Integer vector of length 1 or 2, depending on the number of dimensions of X, with strictly positive entries.}
}
\value{
A numeric vector of length \code{floor(n[1] / l[1]) * floor(n[2] / l[2])}.
}
\description{
This function returns a vector of the block means for a given random field X.
}
\examples{
X <- genField(c(50, 100), H = 100, type = 2)
M <- Mu(X, l = c(10, 20))

plot(X)
image(matrix(M, ncol = 5))

}
