% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mu}
\alias{Mu}
\title{Mu}
\usage{
Mu(X, l)
}
\arguments{
\item{X}{Numeric vector or matrix.}

\item{l}{block length. Numeric vector of the same length as X has dimensions.}
}
\value{
A numeric vector.
}
\description{
This function returns a vector of the block means for a given random field X.
}
