% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mu}
\alias{Mu}
\title{Mu}
\usage{
Mu(X, l)
}
\arguments{
\item{X}{Numeric vector or matrix.}

\item{l}{block length. Numeric vector of length 1 or 2, depending on the number of dimensions of X.}
}
\value{
A numeric vector of length \code{floor(n[1] / l[1]) * floor(n[2] / l[2])}.
}
\description{
This function returns a vector of the block means for a given random field X.
}
\examples{
X <- genField(c(50, 100), H = 100, type = 2)
M <- Mu(X, c(10, 20))

plot(X)
image(matrix(M, ncol = 5))

}
