% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{plot.RandomField}
\alias{plot.RandomField}
\title{Plot random field}
\usage{
\method{plot}{RandomField}(x, main = "", colors, name = "value", alpha = 1, p = NULL, ...)
}
\arguments{
\item{x}{\code{RandomField} object.}

\item{main}{title to the plot.}

\item{colors}{vector of colors over which a gradient is created.}

\item{name}{title of the legend, character string.}

\item{alpha}{opacity of the raster tiles. Numeric value between 0 and 1.}

\item{p}{additional ggplot2 component added BEFORE this function's components.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
a ggplot2 object. Function is mostly called for its side effect.
}
\description{
Plot random field
}
\examples{
plot(genField(c(50, 50)))
plot(genField(c(50, 50), type = 2))

}
