% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SChangeBlock-package.R
\docType{package}
\name{SChangeBlock-package}
\alias{SChangeBlock}
\alias{SChangeBlock-package}
\title{SChangeBlock: Spatial Structural Change Detection by an Analysis of Variability Between Blocks of Observations}
\description{
Provides methods to detect structural changes in time series or random fields (spatial data). Focus is on the detection of abrupt changes or trends in independent data, but the package also provides a function to de-correlate data with dependence. The functions are based on the test suggested in Schmidt (2024) \url{doi:10.3150/23-BEJ1686} and the work in Görz and Fried (2025+).
}
\author{
\strong{Maintainer}: Sheila Goerz \email{sheila.goerz@tu-dortmund.de}

Other contributors:
\itemize{
  \item Roland Fried \email{msnat@statistik.tu-dortmund.de} [contributor, thesis advisor]
}

}
\keyword{internal}
